+++ tomcat/src/main/java/nextstep/jwp/presentation/StaticResourceController.java
@@ -1,8 +1,9 @@
 package nextstep.jwp.presentation;
 
 import java.nio.file.Path;
-import org.apache.coyote.HttpRequest;
-import org.apache.coyote.HttpResponse;
+import org.apache.coyote.http11.common.HttpRequest;
+import org.apache.coyote.http11.common.HttpResponse;
+import org.apache.coyote.http11.util.StaticResource;
 
 public class StaticResourceController extends AbstractController {
