+++ tomcat/src/main/java/nextstep/jwp/service/UserService.java
@@ -6,7 +6,7 @@ import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UnauthorizedException;
 import nextstep.jwp.exception.UserNotFoundException;
 import nextstep.jwp.model.User;
-import org.apache.coyote.HttpRequest;
+import org.apache.coyote.http11.common.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
