+++ tomcat/src/main/java/org/apache/coyote/http11/common/HttpHeaders.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package org.apache.coyote.http11.common;
 
 import java.util.HashMap;
 import java.util.List;
