+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/common/HttpRequestTest.java
@@ -1,11 +1,11 @@
-package nextstep.org.apache.coyote;
+package nextstep.org.apache.coyote.http11.common;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.apache.coyote.HttpRequest;
+import org.apache.coyote.http11.common.HttpRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
